% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorMatCol.R
\name{CorMatCol}
\alias{CorMatCol}
\title{Coloured Corrlation Matrix}
\usage{
CorMatCol(data, cor.var, col.var = NULL, cor.names = NULL,
  col.names = NULL, palette = NULL, alpha = 0.8, method = NULL,
  pearson = TRUE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{cor.var}{A list of column names of the variables to include in the correlation matrix.}

\item{col.var}{An optional column name of the variable to be mapped to the colour aesthetic. Can be either continuous or discrete.}

\item{cor.names}{An optional list of names for axis labels of the variables in the correlation matrix.}

\item{col.names}{An optional names for the coloured variable legend.}

\item{palette}{An optional vector of colours to use as a colour palette for the variable mapped to the colour aesthetic.}

\item{alpha}{Trasperancy value of the points in the scatterplot. Defaults to .8.}

\item{method}{An optional variable to add geom_smooth() to the scatterplot. Inherits from the "method" argument of geom_smooth().}

\item{pearson}{If TRUE (default) writes the coefficients from Pearson Product-Moment Correlations between each two variables above the diagonal. If FALSE leaves the area above the diagonal blank.}
}
\description{
This function creates a matrix of plots, given a dataframe and column names. The diagonal is histograms of each of the selected variables, whereas the area under the diagonal consists of scatterplots. Additionally, the function can accept a parameter to map to a colour aesthetic in the scatterplots, can add geom_smooth to the scatterplots, and can calculate Pearson Product-Moment correlation coefficients, written above the diagonal.
}
\examples{
CorMatCol(data = iris, cor.var = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"), col.var = "Species")
}
